+++ demisto_sdk/commands/format/tests/test_formatting_yml_test.py
@@ -23,13 +23,15 @@ from demisto_sdk.tests.constants_test import (
     DESTINATION_FORMAT_PLAYBOOK, DESTINATION_FORMAT_PLAYBOOK_COPY,
     DESTINATION_FORMAT_SCRIPT_COPY, DESTINATION_FORMAT_TEST_PLAYBOOK,
     EQUAL_VAL_FORMAT_PLAYBOOK_DESTINATION, EQUAL_VAL_FORMAT_PLAYBOOK_SOURCE,
-    EQUAL_VAL_PATH, FEED_INTEGRATION_INVALID, FEED_INTEGRATION_VALID, GIT_ROOT,
-    INTEGRATION_PATH, PLAYBOOK_PATH, SOURCE_FORMAT_INTEGRATION_COPY,
-    SOURCE_FORMAT_INTEGRATION_INVALID, SOURCE_FORMAT_INTEGRATION_VALID,
-    SOURCE_FORMAT_PLAYBOOK, SOURCE_FORMAT_PLAYBOOK_COPY,
-    SOURCE_FORMAT_SCRIPT_COPY, SOURCE_FORMAT_TEST_PLAYBOOK, TEST_PLAYBOOK_PATH)
+    EQUAL_VAL_PATH, FEED_INTEGRATION_EMPTY_VALID, FEED_INTEGRATION_INVALID,
+    FEED_INTEGRATION_VALID, GIT_ROOT, INTEGRATION_PATH, PLAYBOOK_PATH,
+    SOURCE_FORMAT_INTEGRATION_COPY, SOURCE_FORMAT_INTEGRATION_INVALID,
+    SOURCE_FORMAT_INTEGRATION_VALID, SOURCE_FORMAT_PLAYBOOK,
+    SOURCE_FORMAT_PLAYBOOK_COPY, SOURCE_FORMAT_SCRIPT_COPY,
+    SOURCE_FORMAT_TEST_PLAYBOOK, TEST_PLAYBOOK_PATH)
 from mock import Mock, patch
 from ruamel.yaml import YAML
+from TestSuite.test_tools import ChangeCWD
 
 ryaml = YAML()
 ryaml.preserve_quotes = True
@@ -477,7 +479,12 @@ class TestFormatting:
             for counter, param in enumerate(params):
                 if 'defaultvalue' in param and param['name'] != 'feed':
                     params[counter].pop('defaultvalue')
-            for param in FEED_REQUIRED_PARAMS:
+                if 'hidden' in param:
+                    param.pop('hidden')
+            for param_details in FEED_REQUIRED_PARAMS:
+                param = {'name': param_details.get('name')}
+                param.update(param_details.get('must_equal', dict()))
+                param.update(param_details.get('must_contain', dict()))
                 assert param in params
         os.remove(target)
         os.rmdir(path)
@@ -498,6 +505,26 @@ class TestFormatting:
         base_yml.set_feed_params_in_config()
         configuration_params = base_yml.data.get('configuration', [])
         assert 'defaultvalue' in configuration_params[0]
+
+    def test_format_on_feed_integration_adds_feed_parameters(self):
+        """
+        Given
+        - Feed integration yml without feed parameters configured.
+
+        When
+        - Running the format command.
+
+        Then
+        - Ensures the feed parameters are added.
+        """
+        base_yml = IntegrationYMLFormat(FEED_INTEGRATION_EMPTY_VALID, path="schema_path", verbose=True)
+        base_yml.set_feed_params_in_config()
+        configuration_params = base_yml.data.get('configuration', [])
+        for param_details in FEED_REQUIRED_PARAMS:
+            param = {'name': param_details.get('name')}
+            param.update(param_details.get('must_equal', dict()))
+            param.update(param_details.get('must_contain', dict()))
+            assert param in configuration_params
 
     def test_set_fetch_params_in_config_with_default_value(self):
         """
@@ -769,3 +796,53 @@ class TestFormatting:
         assert is_string_uuid(playbook_yml.data['tasks']['2']['task']['id']) and \
             is_string_uuid(playbook_yml.data['tasks']['2']['taskid'])
         assert playbook_yml.data['tasks']['2']['task']['id'] == playbook_yml.data['tasks']['2']['taskid']
+
+    def test_check_for_subplaybook_usages(self, repo):
+        """
+        Given
+            - A test playbook file
+        When
+            - Run check_for_subplaybook_usages command
+        Then
+            - Ensure that the subplaybook id is replaced from the uuid to the playbook name.
+        """
+        pack = repo.create_pack('pack')
+        playbook = pack.create_playbook('LargePlaybook')
+        test_task = {
+            "id": "1",
+            "ignoreworker": False,
+            "isautoswitchedtoquietmode": False,
+            "isoversize": False,
+            "nexttasks": {
+                '#none#': ["3"]
+            },
+            "note": False,
+            "quietmode": 0,
+            "separatecontext": True,
+            "skipunavailable": False,
+            "task": {
+                "brand": "",
+                "id": "dcf48154-7e80-42b3-8464-7156e1cd3d10",
+                "iscommand": False,
+                "name": "my-sub-playbook",
+                "playbookId": "03d4f06c-ad13-47dd-8955-c8f7ccd5cba1",
+                "type": "playbook",
+                "version": -1
+            },
+            "taskid": "dcf48154-7e80-42b3-8464-7156e1cd3d10",
+            "timertriggers": [],
+            "type": "playbook"
+        }
+        playbook.create_default_playbook()
+        playbook_data = playbook.yml.read_dict()
+        playbook_data['tasks']['1'] = test_task
+        playbook.yml.write_dict(playbook_data)
+        playbook_yml = PlaybookYMLFormat(SOURCE_FORMAT_PLAYBOOK_COPY, path='', verbose=True)
+
+        with ChangeCWD(repo.path):
+            playbook_yml.check_for_subplaybook_usages(file_path=playbook.yml.rel_path,
+                                                      current_playbook_id="03d4f06c-ad13-47dd-8955-c8f7ccd5cba1",
+                                                      new_playbook_id="my-sub-playbook")
+
+        playbook_data = playbook.yml.read_dict()
+        assert playbook_data['tasks']['1']['task']['playbookId'] == "my-sub-playbook"